// web/src/api/model/manufacturerResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * nimap API
 * OpenAPI spec version: dev
 */

export interface ManufacturerResponse {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  createdAt: string;
  id: string;
  name: string;
  notes?: string;
  updatedAt: string;
  website?: string;
}